backend/pyproject.toml
[project]
name = "dah-open-rag"
version = "0.1.0"
description = "Modular open-source RAG system for African data — Data Afrique Hub"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.8",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.0.0",
    "numpy>=2.4.2",
    "sentence-transformers>=5.2.2",
    "chromadb>=1.5.0",
    "faiss-cpu>=1.13.2",
    "mistralai>=1.12.1",
    "cohere>=5.20.4",
    "ollama>=0.6.1",
    "huggingface-hub>=1.4.1",
    "pypdf>=6.7.0",
    "langdetect>=1.0.9",
    "tqdm>=4.67.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
